refactor: move TransformRule/TransformSet to homeboy-refactor-contract
TransformRule (a find/replace rule) and TransformSet (a collection of them) are
pure serde config types — the input declared by extensions for test-drift
transforms — that lived in the 43k-LOC refactor::transform engine module. Move
the two types (+ their serde default fns default_files_glob/default_context) into
homeboy-refactor-contract alongside the autofix-result types (#8774); the 52-fn
transform engine stays in core.

- refactor::transform re-exports TransformRule/TransformSet so the engine and
  crate::refactor::{TransformRule, TransformSet} call sites are unchanged.
- extension test/drift.rs (which constructs TransformRule) now imports it from
  homeboy_refactor_contract directly and is FULLY severed from crate::refactor
  (was the only thing it used). test/workflow.rs takes TransformSet from the
  contract too, keeping only its genuine autofix-engine dep.

Extension's refactor edge drops from 3 files to 2. Continues draining the
extension subsystem's type edges into the keystone contracts.

Also removes a pre-existing duplicate candidate_ref field in two agent_task_
promotion test constructors (two fixes collided on main, producing a duplicate-
field E0062 that made cargo check --workspace --tests red on clean main).
Verified the dup is on origin/main and not introduced here.

Refs #8425
